ref(sdk): rename joint apis from "injectXIds" to "injectNewXIds"
