+++ packages/core/src/api/index.ts
@@ -8,9 +8,9 @@ export * from './composed/spore/destroySpore';
 // Joint APIs
 export * from './joints/cluster/injectNewClusterOutput';
 export * from './joints/cluster/injectLiveClusterCell';
-export * from './joints/cluster/injectClusterIds';
+export * from './joints/cluster/injectNewClusterIds';
 export * from './joints/cluster/getClusterCell';
 export * from './joints/spore/injectNewSporeOutput';
 export * from './joints/spore/injectLiveSporeCell';
-export * from './joints/spore/injectSporeIds';
+export * from './joints/spore/injectNewSporeIds';
 export * from './joints/spore/getSporeCell';
