+++ packages/core/src/api/joints/spore/injectNewSporeIds.ts
@@ -2,9 +2,9 @@ import { helpers } from '@ckb-lumos/lumos';
 import { generateTypeIdsByOutputs } from '../../../helpers';
 import { getSporeConfig, getSporeScript, isSporeScriptSupported, SporeConfig } from '../../../config';
 
-export function injectClusterIds(props: {
+export function injectNewSporeIds(props: {
   txSkeleton: helpers.TransactionSkeletonType;
-  outputIndices?: number[];
+  outputIndices: number[];
   config?: SporeConfig;
 }): helpers.TransactionSkeletonType {
   // Env
@@ -17,34 +17,33 @@ export function injectClusterIds(props: {
   const inputs = txSkeleton.get('inputs');
   const firstInput = inputs.get(0);
   if (!firstInput) {
-    throw new Error('Cannot generate Cluster Id because Transaction.inputs[0] does not exist');
+    throw new Error('Cannot generate Spore Id because Transaction.inputs[0] does not exist');
   }
 
-  // Get ClusterType script
-  const clusterScript = getSporeScript(config, 'Cluster');
+  // Get SporeType script
+  const sporeScript = getSporeScript(config, 'Spore');
 
   // Calculates TypeIds by the outputs' indices
   let outputs = txSkeleton.get('outputs');
   let typeIdGroup = generateTypeIdsByOutputs(firstInput, outputs.toArray(), (cell) => {
-    return !!cell.cellOutput.type && isSporeScriptSupported(clusterScript, cell.cellOutput.type);
+    return !!cell.cellOutput.type && isSporeScriptSupported(sporeScript, cell.cellOutput.type);
   });
 
-  // If `clusterOutputIndices` is provided, filter the result
+  // If `sporeOutputIndices` is provided, filter the result
   if (props.outputIndices) {
     typeIdGroup = typeIdGroup.filter(([outputIndex]) => {
       const index = props.outputIndices!.findIndex((index) => index === outputIndex);
       return index >= 0;
     });
     if (typeIdGroup.length !== props.outputIndices.length) {
-      throw new Error('Cannot generate Cluster Id because clusterOutputIndices cannot be fully handled');
+      throw new Error('Cannot generate Spore Id because sporeOutputIndices cannot be fully handled');
     }
   }
 
-  // Update results
   for (const [index, typeId] of typeIdGroup) {
     const output = outputs.get(index);
     if (!output) {
-      throw new Error(`Cannot generate Cluster Id because Transaction.outputs[${index}] does not exist`);
+      throw new Error(`Cannot generate Spore Id because Transaction.outputs[${index}] does not exist`);
     }
 
     output.cellOutput.type!.args = typeId;
